requirements.txt
flask==2.0.1
flask-wtf==0.15.1
flask-login==0.5.0
python-dotenv==0.19.0
watchdog==2.1.5
psutil==5.8.0
requests==2.26.0
werkzeug==2.0.1